chore: extract cross-language test case
